requirements.txt
grpcio
grpcio-tools
rasterio==1.4.2
torch==2.2.0
torchvision==0.17.0
matplotlib==3.9.0
tifffile==2022.10.10
numpy==1.26.4
pillow==10.4.0
dill==0.3.8
geopandas==1.0.0
pytest
dash==2.18.2
plotly==5.14.0
pandas==2.2.3
dash_bootstrap_components==1.4.1
dash_leaflet==0.1.23
dash_extensions==1.0.0
owslib